+++ jammming/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    testEnvironment: 'jsdom',
+    setupFilesAfterEnv: ['jammming/src/test/App.test.js'],
+};
+
