--- jammming/jest.config.json
@@ -1,10 +1,0 @@
-{
-    "testMatch": ["**/test/**/*.test.js"],
-    "transform": {
-      "^.+\\.(js|jsx)$": "babel-jest"
-    },
-    "moduleFileExtensions": ["js", "jsx"],
-    "setupFilesAfterEnv": ["@testing-library/jest-dom/extend-expect"],
-    "testEnvironment": "jsdom"
-}
-
